Use pint parameters and axes everywhere
This commit introduces pint into this project. Now all parameters and
axes are formatted using pint (where possible) to facilitate plotting
and readability of the codebase. Furher, units provide an improved
debugging experience by raising exceptions when a developer attempts
to construct a model that results in non-physical units.

This commit is inline with commit 551d072 in the fitkit project where
parameters may be defined using units for more logical model
construction.

Pint is not full compatible with numpy or sympy and as such some hacks
have been necessary to get this to work. Hopefully these hacks can be
removed with future improvements to pint. For example, see `in_hz` in
qutilities/notch.py for a situation where a 'pint hack' was required.