Update reflection fitting algorithm. Now it works
This commit includes some of the latest UI changes from fitkit and some
new maths to help find a starting point for the fitting parameters. I
have tested this on real experimental data so I'm confident that its
working ok now. Before I just wrote the implementation without anything
to test it on.