Improve hybrid resonance model
Changin the parameter f_b to df simplifies interacting with the model as
it represents the detuning between the cavity and the cavity + drive
frequencies. Changing 'df' has the effect of translating the hybrid
peaks in frequency, where as before changing f_b had a non-trivial
effect on the magnitude response.

This change also specifies the complex coupling quality factor phase in
terms of radians to be consistent with the other resonance models in
qutilities.