Relabel parameters and update for newer fitkit
Parameter labels have changed so as to be consistent with other
resonance models in qutilities.

The code has also been updated to reflect recent changes in fitkit and
now uses xarray.DataArray as the call data type. The new call type is
advantageous because it allows the user to store metadata in the NetCDF
files it can write to. In particular, we store the parameter values used
ot generate data from the reflection model.