Fix reflection model for DPA
In this commit I make some style improvments, but the significant change
is that I remove 'lambda' from the numerator of the reflection
expression. This misplaced component comes from the idler and is offset
in frequency if w != wp/2. This makes a huge difference to the fits. I
have checked that if lambda -> 0 we get the standard reflection model
back.